Убрал ненужный класс, упорядочил импорты
